Moving tests to integration
